# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    name: ERG Support
    url: https://erg.ninja/support
  description: The unified backend API for ERG — a multi-tenant eLearning, content
    management, and automation platform.
  title: ERG Platform API
  version: "1.0"
paths:
...
  /api/access-control/bulk-assign-role:
    post:
      consumes:
      - application/json
      description: Assigns a role to multiple users.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Bulk assign role
      tags:
      - Access Control
  /api/access-control/feature-config:
    get:
      description: Returns feature flags/config for access control.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get feature configuration
      tags:
      - Access Control
  /api/access-control/permission-groups:
    get:
      description: Returns permission groups for categorization.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List permission groups
      tags:
      - Access Control
  /api/access-control/permissions:
    get:
      description: Returns all system permissions.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all permissions
      tags:
      - Access Control
  /api/access-control/roles:
    get:
      description: Returns all available roles.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all roles
      tags:
      - Access Control
    post:
      consumes:
      - application/json
      description: Creates a new role with permissions.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a role
      tags:
      - Access Control
  /api/access-control/roles/{id}:
    delete:
      description: Deletes a role by ID.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - Access Control
    get:
      description: Returns a role by ID.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a role
      tags:
      - Access Control
    put:
      consumes:
      - application/json
      description: Updates an existing role.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a role
      tags:
      - Access Control
  /api/access-control/users/{userId}/effective-permissions:
    get:
      description: Returns resolved permissions for a user (roles + overrides).
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get effective permissions
      tags:
      - Access Control
  /api/access-control/users/{userId}/permission-overrides:
    get:
      description: Returns permission overrides for a specific user.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List permission overrides
      tags:
      - Access Control
  /api/access-control/users/{userId}/permissions:
    post:
      consumes:
      - application/json
      description: Adds a permission override for a user.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add permission override
      tags:
      - Access Control
  /api/access-control/users/{userId}/permissions/{overrideId}:
    delete:
      description: Removes a permission override.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Override ID
        in: path
        name: overrideId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete permission override
      tags:
      - Access Control
  /api/access-control/users/{userId}/preview:
    post:
      consumes:
      - application/json
      description: Preview what changes would look like before applying.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Preview permission changes
      tags:
      - Access Control
  /api/access-control/users/{userId}/roles:
    post:
      consumes:
      - application/json
      description: Assigns one or more roles to a specific user.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Assign roles to user
      tags:
      - Access Control
  /api/access-control/users/{userId}/roles/{roleId}:
    delete:
      description: Removes a specific role from a user.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove role from user
      tags:
      - Access Control
...
  /api/users/{userID}/roles:
    post:
      consumes:
      - application/json
      description: Assigns one or more roles to a user.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: Assign roles to user'
      tags:
      - Users Admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"